// src/models/enums/paymentData.ts
export enum PaystackStatus {
  SUCCESS = 'success',
  FAILED = 'failed',
  PENDING = 'pending',
}

export enum PaystackChargeStatus {
  FAILED = 'failed',
  SUCCESS = 'success',
  SEND_OTP = 'send_otp',
  PENDING = 'pending',
  SEND_PIN = 'send_pin',
  OPEN_URL = 'open_url',
  SEND_PHONE = 'send_phone',
  SEND_BIRTHDAY = 'send_birthday',
  PAY_OFFLINE = 'pay_offline',
  SEND_ADDRESS = 'send_address',
}